+++ test/form/toy_tracker.cpp
@@ -10,9 +10,9 @@ std::vector<TrackStart> ToyTracker::operator()()
   int32_t const npx = generateRandom() % m_maxTracks;
   std::vector<TrackStart> points(npx);
   for (int nelement = 0; nelement < npx; ++nelement) {
-    points[nelement] = TrackStart(float(generateRandom()) / random_max,
-                                  float(generateRandom()) / random_max,
-                                  float(generateRandom()) / random_max);
+    points[nelement] = TrackStart((float)generateRandom() / (float)random_max,
+                                  (float)generateRandom() / (float)random_max,
+                                  (float)generateRandom() / (float)random_max);
   }
 
   return points;
